generall cleaning up and refactoring mapping
