+++ project-staging.yaml
@@ -1,15 +1,27 @@
-specVersion: 0.2.0
+specVersion: 1.0.0
 name: turing-staging-subql
 version: 0.0.4
 description: Any queries for subquery for Turing Staging network
 repository: https://github.com/oak/oak-subql
+
+runner:
+  node:
+    name: "@subql/node"
+    version: "*"
+  query:
+    name: "@subql/query"
+    version: "*"
+
 schema:
   file: ./schema.graphql
+
 network:
   endpoint: 'wss://rpc.turing-staging.oak.tech'
-  genesisHash: '0xd54f0988402deb4548538626ce37e4a318441ea0529ca369400ebec4e04dfe4b'
+  chainId: '0xd54f0988402deb4548538626ce37e4a318441ea0529ca369400ebec4e04dfe4b'
+  #dictionary: "https://api.subquery.network/sq/subquery/kusama-dictionary"
   chaintypes:
     file: ./dist/chaintypes.js
+
 dataSources:
   - kind: substrate/Runtime
     startBlock: 1
@@ -20,3 +32,5 @@ dataSources:
           kind: substrate/BlockHandler
         - handler: handleEvent
           kind: substrate/EventHandler
+        - handler: handleCall
+          kind: substrate/CallHandler
