+++ src/index.ts
@@ -1,4 +1,4 @@
 //Exports all handler functions
 import '@polkadot/api-augment';
 export * from './mappings/mappingHandlers'
-export * from './mappings/eventHandlers'
+//export * from './mappings/eventHandlers'
